Mejora en Creacion de Empresa y Area Comercial
- Ajustes en Roles y Permisos

+++ database/migrations/2025_01_29_221428_alter_ciudades_table.php
@@ -19,7 +19,7 @@ class AlterCiudadesTable extends Migration
         });
 
         Schema::table('ciudades', function (Blueprint $table) {
-            $table->unsignedInteger('departamento_id')->change();
+            $table->unsignedBigInteger('departamento_id')->change();
             $table->foreign('departamento_id')->references('id')->on('departamentos');
 
             $table->renameColumn('ciudad', 'nombre');
